codec2: CB2 fix dlclose crash. [1/1]
PD#SWPL-122913

Problem:
Using ffmpeg compiled by bp, when the playback exits,
the dlclose function of codec2 cause crash.

Solution:
Execute dlopen() when opening the soft video decode for the
first time, and it will not dlclose() after exiting the playback,
that is, the first handle will be used every time.

Verify:
AndroidT

Change-Id: I686bd67adbc10d19e0f9f1f8b5ac650eeb570f9c
Signed-off-by: kenan.liu <kenan.liu@amlogic.com>